// ServerlessFramework/src/enums/workflows.ts
export enum WorkflowEnum {
  newSimESIM = 'newSimESIM',
  newEsimCollectionPoint = 'newEsimCollectionPoint',
  unlimitedPlans = 'Unlimited Plans',
  activation = 'Activate',
  checkBalance = 'Check Balance',
  topUp = 'Top - Up SIM',
  shopify = 'Shopify New Sim',
  Location = 'location',
  airport = 'Airport',
}

export enum ParentWorkflow {
  Buy = 'buy',
  Activation = 'activation',
  Balance = 'balance',
  Recharge = 'recharge',
  Payment = 'payment',
  Location = 'location',
}

export enum CRMWorkFlow {
  Buy = 'buy',
  Activation = 'activation',
  Balance = 'balance',
  Recharge = 'recharge',
  Payment = 'payment',
  Cds = 'cds', // Crm update flow
  Location = 'location',
  parcelNinja = 'parcelNinja',
  Account = 'account',
  SimPurchase = 'SimPurchase',
  Association = 'Association',
  PartnerTerm = 'PartnerTerm',
  Commission = 'commission',
}